Refactor: Base de données

+++ src/main/java/fr/cytech/projetjeespring/models/Payslip.java
@@ -2,6 +2,8 @@ package fr.cytech.projetjeespring.models;
 
 import fr.cytech.projetjeespring.entities.Employee;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -13,13 +15,13 @@ public class Payslip {
     private LocalDate endDate;
     private int monthsCount;
 
-    private float monthlyGrossSalary;
-    private float totalGrossSalary;
-    private float totalBonuses;
+    private BigDecimal monthlyGrossSalary;
+    private BigDecimal totalGrossSalary;
+    private BigDecimal totalBonuses;
 
     private List<String> bonusBreakdown;
-    private Map<String, Float> deductionBreakdown;
+    private Map<String, BigDecimal> deductionBreakdown;
 
-    private float totalDeductions;
-    private float netPay;
+    private BigDecimal totalDeductions;
+    private BigDecimal netPay;
 }
